requirements.txt
addict                      # MIT
cerberus                    # ISC
cotyledon                   # Apache License 2.0
flask                       # BSD 3-Clause License
flask_restplus              # BSD 3-Clause License
honcho                      # MIT
kubernetes                  # Apache License 2.0
neo4j                       # Apache License 2.0
python-arango               # MIT
pyyaml                      # MIT
requests                    # Apache License 2.0
